chore: split workspace impl as module
