+++ libs/jwst/src/workspaces/plugins/mod.rs
@@ -2,10 +2,10 @@
 mod indexing;
 mod plugin;
 
-use super::*;
+use super::{metadata::SEARCH_INDEX, *};
 
 #[cfg(feature = "workspace-search")]
-pub(super) use indexing::IndexingPluginImpl;
+use indexing::IndexingPluginImpl;
 pub(super) use plugin::{PluginImpl, PluginMap, PluginRegister};
 
 #[cfg(feature = "workspace-search")]
@@ -24,7 +24,7 @@ fn insert_plugin(
 }
 
 /// Setup plugin: [indexing]
-pub(super) fn setup_plugin(workspace: Workspace) -> Workspace {
+pub(crate) fn setup_plugin(workspace: Workspace) -> Workspace {
     // default plugins
     if cfg!(feature = "workspace-search") {
         // Set up indexing plugin
@@ -32,5 +32,57 @@ pub(super) fn setup_plugin(workspace: Workspace) -> Workspace {
             .expect("Failed to setup search plugin")
     } else {
         workspace
+    }
+}
+
+impl Workspace {
+    /// Allow the plugin to run any necessary updates it could have flagged via observers.
+    /// See [plugins].
+    fn update_plugin<P: PluginImpl>(&self) -> Result<(), Box<dyn std::error::Error>> {
+        self.plugins.update_plugin::<P>(self)
+    }
+
+    /// See [plugins].
+    fn with_plugin<P: PluginImpl, T>(&self, cb: impl Fn(&P) -> T) -> Option<T> {
+        self.plugins.with_plugin::<P, T>(cb)
+    }
+
+    #[cfg(feature = "workspace-search")]
+    pub fn search<S: AsRef<str>>(
+        &self,
+        query: S,
+    ) -> Result<SearchResults, Box<dyn std::error::Error>> {
+        // refresh index if doc has update
+        self.update_plugin::<IndexingPluginImpl>()?;
+
+        let query = query.as_ref();
+
+        self.with_plugin::<IndexingPluginImpl, Result<SearchResults, Box<dyn std::error::Error>>>(
+            |search_plugin| search_plugin.search(query),
+        )
+        .expect("text search was set up by default")
+    }
+
+    #[cfg(feature = "workspace-search")]
+    pub fn search_result(&self, query: String) -> String {
+        match self.search(query) {
+            Ok(list) => serde_json::to_string(&list).unwrap(),
+            Err(_) => "[]".to_string(),
+        }
+    }
+
+    pub fn set_search_index(&self, fields: Vec<String>) -> JwstResult<bool> {
+        match fields.iter().find(|&field| field.is_empty()) {
+            Some(field) => {
+                error!("field name cannot be empty: {}", field);
+                Ok(false)
+            }
+            None => {
+                let value = serde_json::to_string(&fields).unwrap();
+                self.with_trx(|mut trx| trx.set_metadata(SEARCH_INDEX, value))?;
+                setup_plugin(self.clone());
+                Ok(true)
+            }
+        }
     }
 }
