+++ libs/jwst/src/workspaces/transaction.rs
@@ -1,8 +1,10 @@
+use std::{thread::sleep, time::Duration};
+
 use crate::utils::JS_INT_RANGE;
 
 use super::*;
 use lib0::any::Any;
-use yrs::{Map, ReadTxn, TransactionMut};
+use yrs::{Map, ReadTxn, Transact, TransactionMut};
 
 pub struct WorkspaceTransaction<'a> {
     pub ws: &'a Workspace,
@@ -72,5 +74,50 @@ impl WorkspaceTransaction<'_> {
 
     pub fn commit(&mut self) {
         self.trx.commit();
+    }
+}
+
+impl Workspace {
+    pub fn with_trx<T>(&self, f: impl FnOnce(WorkspaceTransaction) -> T) -> T {
+        let doc = self.doc();
+        let trx = WorkspaceTransaction {
+            trx: doc.transact_mut(),
+            ws: self,
+        };
+
+        f(trx)
+    }
+
+    pub fn try_with_trx<T>(&self, f: impl FnOnce(WorkspaceTransaction) -> T) -> Option<T> {
+        match self.doc().try_transact_mut() {
+            Ok(trx) => {
+                let trx = WorkspaceTransaction { trx, ws: self };
+                Some(f(trx))
+            }
+            Err(e) => {
+                info!("try_with_trx error: {}", e);
+                None
+            }
+        }
+    }
+
+    pub fn retry_with_trx<T>(
+        &self,
+        f: impl FnOnce(WorkspaceTransaction) -> T,
+        mut retry: i32,
+    ) -> Option<T> {
+        let trx = loop {
+            if let Ok(trx) = self.doc.try_transact_mut() {
+                break trx;
+            } else if retry > 0 {
+                retry -= 1;
+                sleep(Duration::from_micros(10));
+            } else {
+                info!("retry_with_trx error");
+                return None;
+            }
+        };
+
+        Some(f(WorkspaceTransaction { trx, ws: self }))
     }
 }
